// src/types/beneficiary.ts
export enum BeneficiaryCategory {
  YETIM_AILESI = 'YETIM_AILESI',
  MULTECI_AILE = 'MULTECI_AILE',
  IHTIYAC_SAHIBI_AILE = 'IHTIYAC_SAHIBI_AILE',
  YETIM_COCUK = 'YETIM_COCUK',
  MULTECI_COCUK = 'MULTECI_COCUK',
  IHTIYAC_SAHIBI_COCUK = 'IHTIYAC_SAHIBI_COCUK',
  YETIM_GENCLIK = 'YETIM_GENCLIK',
  MULTECI_GENCLIK = 'MULTECI_GENCLIK',
  IHTIYAC_SAHIBI_GENCLIK = 'IHTIYAC_SAHIBI_GENCLIK',
}

export enum FundRegion {
  AVRUPA = 'AVRUPA',
  SERBEST = 'SERBEST',
}